requirements.txt
nltk==3.2.5
python-dateutil==2.7.2
tokenize-uk==0.2.0
tqdm==4.23.4